[BE] FIX: Ban 관련 메서드 Policy로 이전 #1038
