+++ backend/src/main/java/org/ftclub/cabinet/user/domain/BanPolicyImpl.java
@@ -1,9 +1,11 @@
 package org.ftclub.cabinet.user.domain;
 
 import java.util.Date;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.ftclub.cabinet.cabinet.domain.LentType;
 import org.ftclub.cabinet.config.CabinetProperties;
+import org.ftclub.cabinet.user.repository.BanHistoryRepository;
 import org.ftclub.cabinet.utils.DateUtil;
 import org.springframework.stereotype.Component;
 
@@ -12,6 +14,7 @@ import org.springframework.stereotype.Component;
 public class BanPolicyImpl implements BanPolicy {
 
 	private final CabinetProperties cabinetProperties;
+	private final BanHistoryRepository banHistoryRepository;
 
 	@Override
 	public BanType verifyForBanType(LentType lentType, Date startAt, Date endedAt, Date expiredAt) {
@@ -28,12 +31,13 @@ public class BanPolicyImpl implements BanPolicy {
 	}
 
 	@Override
-	public Date getBanDate(BanType banType, Date endedAt, Date expiredAt) {
+	public Date getBanDate(BanType banType, Date endedAt, Date expiredAt, Long userId) {
 		if (banType == BanType.SHARE) {
 			return DateUtil.addDaysToDate(endedAt, cabinetProperties.getPenaltyDayShare());
 		} else {
-			int currentBan = DateUtil.calculateTwoDateDiffAbs(endedAt, expiredAt).intValue();
-			return DateUtil.addDaysToDate(endedAt, currentBan);
+			int currentBanDays = DateUtil.calculateTwoDateDiffAbs(endedAt, expiredAt).intValue();
+			int accumulateBanDays = getAccumulateBanDaysByUserId(userId).intValue();
+			return DateUtil.addDaysToDate(endedAt, currentBanDays + accumulateBanDays);
 		}
 	}
 
@@ -45,5 +49,16 @@ public class BanPolicyImpl implements BanPolicy {
 	@Override
 	public boolean isActiveBanHistory(Date unbannedAt, Date now) {
 		return now.before(unbannedAt);
+	}
+
+	@Override
+	public Long getAccumulateBanDaysByUserId(Long userId) {
+		List<BanHistory> banHistories = banHistoryRepository.findBanHistoriesByUserId(userId);
+		Long accumulateDays = 0L;
+		for (BanHistory history : banHistories) {
+			accumulateDays += DateUtil.calculateTwoDateDiffAbs(history.getBannedAt(),
+					history.getUnbannedAt());
+		}
+		return accumulateDays;
 	}
 }
